feat: paste externally copied images onto editor
