+++ frontend/src/components/DropTargetOverlay.vue
@@ -10,12 +10,10 @@
 
 <script setup>
 import { ref, useTemplateRef } from 'vue'
-import { toast } from 'vue-sonner'
-
-import { FileUploadHandler } from 'frappe-ui'
 
 import { presentationId } from '@/stores/presentation'
 import { addMediaElement } from '@/stores/element'
+import { handleUploadedMedia } from '@/utils/mediaUploads'
 
 const emit = defineEmits(['hideOverlay'])
 
@@ -26,46 +24,9 @@ const handleDragLeave = (e) => {
 	emit('hideOverlay', false)
 }
 
-const fileUploadHandler = new FileUploadHandler()
-
-const uploadMedia = (file, fileType) => {
-	return new Promise((resolve, reject) => {
-		fileUploadHandler
-			.upload(file, {
-				doctype: 'Presentation',
-				docname: presentationId.value,
-				private: true,
-			})
-			.then((fileDoc) => {
-				addMediaElement(fileDoc, fileType)
-				resolve(fileDoc)
-			})
-			.catch((error) => {
-				reject(error)
-			})
-	})
-}
-
-const uploadFiles = (files) => {
-	files.forEach((file, index) => {
-		const fileType = file.type.split('/')[0]
-		if (!['image', 'video'].includes(fileType)) return
-
-		const toastProps = {
-			loading: `Uploading (${index + 1}/${files.length}): ${file.name}`,
-			success: (data) => `Uploaded: ${file.name}`,
-			error: (data) => 'Upload failed. Please try again.',
-		}
-
-		// run after current call stack so toast's expand animation works
-		setTimeout(() => toast.promise(uploadMedia(file, fileType), toastProps), 0)
-	})
-}
-
 const handleMediaDrop = async (e) => {
 	e.preventDefault()
 	emit('hideOverlay', false)
-	const files = e.dataTransfer.files
-	uploadFiles(files)
+	handleUploadedMedia(e.dataTransfer.files)
 }
 </script>
